User nicknames as public user IDs: Change UserR route
Making UserR take a Text that refers to a whole new User field is a big
change. Many handlers and templates need to be updated. Therefore please
read below carefully to understand my plan and what exactly this commit
does.

The GOAL is to move all UserId-based routes to use UserHandle instead,
which is Text. That "handle" references a new 'userNick` field in the
User table. But since this is a huge change, this commit does a smaller
change. The rest will come in the next patches. What this commit does
is:

1. Update the UserR route and add a deprecated UserByIdR route
2. Add the 'userNick' field to account creation and account update forms
3. Require a pattern for the 'userNick' field to avoid all-digit
   nicknames and other potential issues
4. Fix all handlers and templates to pass the 'userNick' to 'UserR'
   instead of the 'userId'. That includes inserting 'get404' calls in
   several places just to get the User value, and other helpers that
   will be removed later. They are used just so that the code builds and
   works after applying this patch.

The next patch(es) will update all the other handlers and probably
change some of the helper functions to work with the new 'userNick'
field.

Before this gets committed, some input/review from dev and desig teams
will be great. But don't be pedantic, as the next commits will change
many things and will remove some of the ugly parts used here
temporarily.

I suggest we make this change in 2 steps because it's so big:

1. Change the routes and UI as needed, and do whatever is needed to make the
   code build again
2. When done, refactor and add helpers and put all the chaos back to order